Directories refactoring
