+++ pkg/rancher-ai-ui/components/BubbleButton.vue
