+++ pkg/rancher-ai-ui/components/message/actions/Action.vue
