+++ pkg/rancher-ai-ui/components/message/actions/index.vue
